Validate utreexo node hashes at compile time
Created a new `macros` module at `floresta-common`. Moved here the `bhash` and `impl_error_from` macros.

Then implemented the `acchashes` macro, following a similar approach as `bhash`.
