+++ crates/floresta-common/src/lib.rs
@@ -10,6 +10,8 @@ use miniscript::Descriptor;
 #[cfg(any(feature = "descriptors-std", feature = "descriptors-no-std"))]
 use miniscript::DescriptorPublicKey;
 use sha2::Digest;
+
+pub mod macros;
 pub mod spsc;
 
 use prelude::*;
@@ -147,14 +149,4 @@ pub mod prelude {
     pub use std::sync;
     pub use std::vec;
     pub use std::vec::Vec;
-}
-#[macro_export]
-macro_rules! impl_error_from {
-    ($thing:ty, $from_thing:ty, $field:ident) => {
-        impl From<$from_thing> for $thing {
-            fn from(e: $from_thing) -> Self {
-                <$thing>::$field(e)
-            }
-        }
-    };
 }
